requirements.txt
certifi==2020.6.20
chardet==3.0.4
code2image==0.0.3
idna==2.10
peewee==3.13.3
Pillow==7.2.0
Pygments==2.6.1
pyTelegramBotAPI==3.7.2
requests==2.24.0
six==1.15.0
toml==0.10.1
urllib3==1.25.9
